+++ topydo/commands/ListCommand.py
@@ -16,13 +16,14 @@
 
 from topydo.lib.Config import config
 from topydo.lib.ExpressionCommand import ExpressionCommand
+from topydo.lib.Filter import InstanceFilter
 from topydo.lib.PrettyPrinter import pretty_printer_factory
-from topydo.lib.PrettyPrinterFilter import (PrettyPrinterHideTagFilter,
-                                            PrettyPrinterIndentFilter)
+from topydo.lib.prettyprinters.Format import PrettyPrinterFormatFilter
+from topydo.lib.TodoListBase import InvalidTodoException
 
 
 class ListCommand(ExpressionCommand):
-    def __init__(self, p_args, p_todolist,
+    def __init__(self, p_args, p_todolist, #pragma: no branch
                  p_out=lambda a: None,
                  p_err=lambda a: None,
                  p_prompt=lambda a: None):
@@ -32,6 +33,8 @@ class ListCommand(ExpressionCommand):
         self.printer = None
         self.sort_expression = config().sort_string()
         self.show_all = False
+        self.ids = None
+        self.format = config().list_format()
 
     def _poke_icalendar(self):
         """
@@ -47,7 +50,7 @@ class ListCommand(ExpressionCommand):
         return True
 
     def _process_flags(self):
-        opts, args = self.getopt('f:n:s:x')
+        opts, args = self.getopt('f:F:i:n:s:x')
 
         for opt, value in opts:
             if opt == '-x':
@@ -71,13 +74,43 @@ class ListCommand(ExpressionCommand):
                     self.printer = DotPrinter(self.todolist)
                 else:
                     self.printer = None
+            elif opt == '-F':
+                self.format = value
             elif opt == '-n':
                 try:
                     self.limit = int(value)
                 except ValueError:
                     pass # use default value in configuration
+            elif opt == '-i':
+                self.ids = value.split(',')
+
+                # when a user requests a specific ID, it should always be shown
+                self.show_all = True
 
         self.args = args
+
+    def _filters(self):
+        """
+        Additional filters to select particular todo items given with the -i
+        flag.
+        """
+        filters = super(ListCommand, self)._filters()
+
+        if self.ids:
+            def get_todo(p_id):
+                """
+                Safely obtains a todo item given the user-supplied ID.
+                Returns None if an invalid ID was entered.
+                """
+                try:
+                    return self.todolist.todo(p_id)
+                except InvalidTodoException:
+                    return None
+
+            todos = [get_todo(i) for i in self.ids]
+            filters.append(InstanceFilter(todos))
+
+        return filters
 
     def _print(self):
         """
@@ -90,11 +123,11 @@ class ListCommand(ExpressionCommand):
         if self.printer is None:
             # create a standard printer with some filters
             indent = config().list_indent()
+            final_format = ' ' * indent + self.format
             hidden_tags = config().hidden_tags()
 
             filters = []
-            filters.append(PrettyPrinterIndentFilter(indent))
-            filters.append(PrettyPrinterHideTagFilter(hidden_tags))
+            filters.append(PrettyPrinterFormatFilter(self.todolist, final_format))
 
             self.printer = pretty_printer_factory(self.todolist, filters)
 
@@ -115,7 +148,8 @@ class ListCommand(ExpressionCommand):
         return True
 
     def usage(self):
-        return """ Synopsis: ls [-x] [-s <sort_expression>] [-f <format>] [expression]"""
+        return """Synopsis: ls [-x] [-s <sort_expression>] [-f <output format>]
+[-F <format string>] [expression]"""
 
     def help(self):
         return """\
@@ -137,6 +171,37 @@ When an expression is given, only the todos matching that expression are shown.
                 an 'ical' tag with a unique ID. Completed todo items may be
                 archived.
      * 'json' - Javascript Object Notation (JSON)
+-F : Specify the format of the text ('text' format), which may contain
+     placeholders that may be expanded if the todo has such attribute. If such
+     attribute does not exist, then it expands to an empty string.
+
+         %c: Absolute creation date.
+         %C: Relative creation date.
+         %d: Absolute due date.
+         %D: Relative due date.
+         %h: Relative due and start date (due in 3 days, started 3 days ago)
+         %H: Like %h with creation date.
+         %i: Todo number.
+         %I: Todo number padded with spaces (always 3 characters wide).
+         %k: List of tags separated by spaces (excluding hidden tags).
+         %K: List of all tags separated by spaces.
+         %p: Priority.
+         %s: Todo text.
+         %S: Todo text, truncated such that an item fits on one line.
+         %t: Absolute creation date.
+         %T: Relative creation date.
+         %x: 'x' followed by absolute completion date.
+         %X: 'x' followed by relative completion date.
+         \%: Literal percent sign.
+
+     Conditional characters can be added with blocks surrounded by curly
+     braces, they will only appear when a placeholder expanded to a value.
+
+     E.g. %{(}p{)} will print (C) when the todo item has priority C, or ''
+     (empty string) when an item has no priority set.
+
+     A tab character serves as a marker to start right alignment.
+-i : Comma separated list of todo IDs to print.
 -s : Sort the list according to a sort expression. Defaults to the expression
      in the configuration.
 -x : Show all todos (i.e. do not filter on dependencies or relevance).