+++ topydo/commands/RevertCommand.py
@@ -21,12 +21,12 @@ from topydo.lib import TodoList
 from topydo.lib.Config import config
 
 class RevertCommand(Command):
-    def __init__(self, p_args, p_todolist,
+    def __init__(self, p_args, p_todolist, #pragma: no branch
                  p_out=lambda a: None,
                  p_err=lambda a: None,
                  p_prompt=lambda a: None):
-        super(RevertCommand, self).__init__(
-            p_args, p_todolist, p_out, p_err, p_prompt=lambda a: None)
+        super(RevertCommand, self).__init__(p_args, p_todolist, p_out, p_err,
+                p_prompt)
 
     def execute(self):
         if not super(RevertCommand, self).execute():