+++ topydo/commands/SortCommand.py
@@ -20,7 +20,7 @@ from topydo.lib.Sorter import Sorter
 
 
 class SortCommand(Command):
-    def __init__(self, p_args, p_todolist,
+    def __init__(self, p_args, p_todolist, #pragma: no branch
                  p_out=lambda a: None,
                  p_err=lambda a: None,
                  p_prompt=lambda a: None):